Firmware refactoring Cleanup defines, move sensors to separate folder and headers, cleanup includes, cleanup connection sequence (WIP)
