+++ src/batterymonitor.h
@@ -34,6 +34,13 @@
     #ifndef PIN_BATTERY_LEVEL
         #error Internal ADC enabled without pin! Please select a pin.
     #endif
+    // Wemos D1 Mini has an internal Voltage Divider with R1=220K and R2=100K > this means, 3.3V analogRead input voltage results in 1023.0
+    // Wemos D1 Mini with Wemos BatteryShiled v1.2.0 or higher: BatteryShield with J2 closed, has an additional 130K resistor. So the resulting Voltage Divider is R1=220K+100K=320K and R2=100K > this means, 4.5V analogRead input voltage results in 1023.0
+    // SlimeVR Board can handle max 5V > so analogRead of 5.0V input will result in 1023.0
+    #define batteryADCMultiplier 1.0 / 1023.0 * (320 + BATTERY_SHIELD_RESISTANCE) / 100
+#elif BATTERY_MONITOR == BAT_MCP3021 || BATTERY_MONITOR == BAT_INTERNAL_MCP3021
+  // Default recommended resistors are 9.1k and 5.1k
+  #define batteryADCMultiplier 3.3 / 1023.0 * 14.2 / 9.1
 #endif
 
 class BatteryMonitor
