+++ src/consts.h
@@ -28,9 +28,9 @@
 #define IMU_MPU6500 2
 #define IMU_BNO080 3
 #define IMU_BNO085 4
-#define IMU_BNO086 4
 #define IMU_BNO055 5
 #define IMU_MPU6050 6
+#define IMU_BNO086 7
 
 #define BOARD_SLIMEVR 1
 #define BOARD_SLIMEVR_DEV 2
